functions/requirements.txt
docopts
geddit
pyyaml
dpath
numpy
scipy
